Feature/add to cart

+++ src/main/java/com/spring/boec/entities/Item.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
 public abstract class Item implements Serializable {
 
   @Column(name = "price")
-  protected double price;
+  protected long price;
 
   @Column(name = "stock")
   protected int stock;
